Move the querying with generic groups out of MolOps to not bloat the library.
