+++ Code/GraphMol/GenericGroups/GenericGroups.h
@@ -274,6 +274,12 @@ const static std::map<
         {"NoCarbonRing", Matchers::NoCarbonRingAtomMatcher},
         {"CXX", Matchers::NoCarbonRingAtomMatcher},
 };
+
+// This is an extension of adjustQueryProperties from GraphMol that allows the search of generic groups 
+RDKIT_GENERICGROUPS_EXPORT void adjustQueryPropertiesWithGenericGroups(
+    RWMol &mol, 
+    const AdjustQueryParameters *inParams);
+
 //! returns false if any of the molecule's generic atoms are not satisfied in
 /// the current match
 RDKIT_GENERICGROUPS_EXPORT bool genericAtomMatcher(
