+++ Code/GraphMol/GenericGroups/generic_tests.cpp
@@ -609,3 +609,55 @@ TEST_CASE("generics are compatible with extraFinalChecks",
     }
   }
 }
+
+void testGenericQueries() {
+  BOOST_LOG(rdInfoLog)
+          << "-----------------------\n Testing generic queries: "
+             "Set GenericQueriesFromProperties"
+          << std::endl;
+  auto query = R"CTAB(alkquery.mol
+Generated by WebMolKit
+
+  0  0  0     0  0            999 V3000
+M  V30 BEGIN CTAB
+M  V30 COUNTS 2 1 1 0 1
+M  V30 BEGIN ATOM
+M  V30 1 O 7.4214 -9.6318 0.0000 0
+M  V30 2 * 8.5430 -10.2809 0.0000 0
+M  V30 END ATOM
+M  V30 BEGIN BOND
+M  V30 1 1 1 2
+M  V30 END BOND
+M  V30 BEGIN SGROUP
+M  V30 1 SUP 0 LABEL=ALK ATOMS=(1 2)
+M  V30 END SGROUP
+M  V30 END CTAB
+M  END)CTAB"_ctab;
+
+  TEST_ASSERT(query);
+
+  std::string target_smi{"c1c(OF)cccc1"};
+  std::unique_ptr<RWMol> t{SmilesToMol(target_smi)};
+  TEST_ASSERT(t);
+
+  MolOps::AdjustQueryParameters params;
+  auto m = new RWMol(*query);
+
+  TEST_ASSERT(m);
+
+  GenericGroups::adjustQueryPropertiesWithGenericGroups(*m, &params);
+  SubstructMatchParameters params_match;
+  params_match.useQueryQueryMatches = true;
+  params_match.useGenericMatchers = true;
+  params_match.maxMatches = 1;
+  auto matchVect = SubstructMatch(*t, *m, params_match);
+  TEST_ASSERT(matchVect.size() == 1);
+
+  delete m;
+  m = new RWMol(*query);
+  params.setGenericQueryFromProperties = true;
+  GenericGroups::adjustQueryPropertiesWithGenericGroups(*m, &params);
+  matchVect = SubstructMatch(*t, *m, params_match);
+  TEST_ASSERT(matchVect.size() == 0);
+  delete m;
+}
